[bugfix] using uint8_t based class descriptor size (limited to 255 bytes)

+++ usbhid_descriptor.c
@@ -35,7 +35,7 @@
  * GetDescriptor class level requests, handled by the class-level handler */
 mbed_error_t      usbhid_get_descriptor(uint8_t             iface_id,
                                         uint8_t            *buf,
-                                        uint32_t           *desc_size,
+                                        uint8_t            *desc_size,
                                         uint32_t            usbdci_handler __attribute__((unused)))
 {
     mbed_error_t errcode = MBED_ERROR_NONE;
